:art:  Improve structure / format of the code：認証処理のカスタムフックを作成処理のカスタムフックを作成
