+++ src/components/loading/SectionLoading.tsx
@@ -1,10 +1,17 @@
 import React from "react";
 
-export const SectionLoading = () => {
+export type LoadingProps = {
+  variant?: "primary" | "secondary" | "danger";
+};
+export const SectionLoading = ({ variant = "primary" }: LoadingProps) => {
   return (
     <div className="flex flex-col items-center justify-center gap-4 ">
       <div className="flex justify-center">
-        <div className="w-5 h-5 border-4 border-green-800 rounded-full animate-spin border-t-transparent"></div>
+        <div
+          className={`w-5 h-5 border-4  rounded-full animate-spin border-t-transparent ${
+            variant === "secondary" ? "border-white" : "border-green-800"
+          }`}
+        ></div>
       </div>
     </div>
   );
